requirements/req_dev.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
boto3==1.38.36
botocore==1.38.36
celery==5.5.3
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==45.0.4
daphne==4.2.0
decorator==5.2.1
Django==5.2.3
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
executing==2.2.0
flower==2.0.1
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
kombu==5.5.4
matplotlib-inline==0.1.7
msgpack==1.1.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
s3transfer==0.13.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
